requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# HTTP client
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Worker dependencies
schedule>=1.2.0
PyPDF2>=3.0.0
rich>=13.0.0

# LLM client (for demo endpoint)
chak-ai>=0.1.0

# Zag framework (install from GitHub)
git+https://github.com/zhixiangxue/zag-ai.git#egg=zagpy[retriever,embedder]
